LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Create-url-with-space.any-expected.txt

CONSOLE MESSAGE: Invalid url for WebSocket ws://web platform.test:49001/echo

PASS Create WebSocket - Pass a URL with a space - SYNTAX_ERR should be thrown
